refactor: -axios cleanup and logging improvement fix: - better server shutdown logging
